Create Character replacing.cpp
Write a program that accepts input from the user and creates another string that contains either a u, l, or - for each character of the original string. Use u when the character is uppercase, and use l when the character is lowercase. If the character is neither uppercase nor lowercase, use -. Print the original string first, then print the modified string.