Create Swapping characters.cpp
Write a program that captures input from the user and swaps the letters two at a time in the string. The first two characters change places, the third and fourth characters change places, etc. Assume that the program will only take an even number of characters, otherwise, there will be an error.