Create File Task 1.cpp
Problem
Write a program that reads a text file and returns the number of lines as well as the total number of characters in the file.
The variable path allows for different text files to be sent to your program for testing. The TRY IT button below will send a test file to your program. You should see the following output:
4 line(s)
228 character(s)