Create Files task 2.cpp
Write a program that reads a comma delimited CSV file with four columns and returns the average of each column in the file. Assume that the CSV files used will only contain 3 rows and 4 columns.
The variable path allows for different text files to be sent to your program for testing. The TRY IT button below will send a test file to your program. You should see the following output:
10 8 6 20
Where 10 represents the average of the numbers in the first column, 8 represents the average of the numbers in the second column, etc.